pyproject.toml
[tool.poetry]
name = "prevencao-fraudes"
version = "0.1.0"
description = ""
authors = ["Felipe Lucas Roll <felipelroll@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pandas = "^2.2.2"
nbformat = "^5.10.4"
plotly-express = "^0.4.1"
scikit-learn = "^1.5.1"
imbalanced-learn = "^0.12.3"
numpy = "^2.1.0"
openpyxl = "^3.1.5"
streamlit = "^1.38.0"
streamlit-dynamic-filters = "^0.1.9"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
